Adds a module to fix Godhome menus.
The Godhome menus have a vanilla bug that prevents them from being closed if your Left Mouse button is bound to Attack. The specific menus this bug affects are the Pantheon menus and the individual boss challenge menus.

This bug is due to a line in the HollowKnightInputModule.SendButtonEventToSelectedObject() method, where it checks to see if the AttackAction has a binding equal to Mouse.LeftButton, and if so, prevents the ICancelHandler's cancel method from being called.

This fix adds a module that hooks that method and adds a secondary check that requires that the Attack button was also the button that was pressed.