Update for 1.5.6.1.1179
Had to swap settings due to removal of SerializableDictionary and the
TValue fields in settings using that being removed. Menus and loads
required changes due to new naming in compiler generated types, I might
want to look into getting the compiler generated type via a contains
that's inexact instead of maintaining that. I also removed the mouse
fixes as the beta now takes care of that. The mod itself was swapped
to .NET 4.7.2. FixDashmaster required a field name change due to the
InControl update.